Modularized container, move every scss files to its own component, moved all global scss into index.scss
