+++ app/components/ApplicationsCard/styles.scss
@@ -1,5 +1,5 @@
 
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .card {
   display: flex;
