+++ app/components/DashboardIcons/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .dashboardIconWrapper {
   margin-right: 37px;
