+++ app/components/Header/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 .microservice-header {
   width: 100%;
   @include centerWithFlex(center);
