+++ app/components/Occupied/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 // MAIN DASHBOARD
 .dashboardArea {
