+++ app/components/Setting/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss'; // Update the import path according to your project structure
+@import '../../stylesheets/index.scss'; 
 
 .settings {
   display: flex;
