+++ app/components/Splash/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss'; // Update the import path according to your project structure
+@import '../../stylesheets/index.scss'; // Update the import path according to your project structure
 
 #splash {
   display: flex;
