+++ app/components/WindowBar/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss'; 
+@import '../../stylesheets/index.scss'; 
 #titleBar {
   position: fixed;
   display: flex;
