+++ app/containers/AWSGraphsContainer/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .AWS-container {
   margin-top: 40px;
