+++ app/containers/MainContainer/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .main-container {
   background-color: $grey;
