+++ app/containers/ModifyMetricsContainer/styles.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .metricsSelector {
     max-width: 90%;
