+++ app/containers/SidebarContainer/SidebarContainer.tsx
@@ -6,19 +6,18 @@ import InfoIcon from '@material-ui/icons/Info';
 import ContactSupportIcon from '@material-ui/icons/ContactSupport';
 import SettingsIcon from '@material-ui/icons/Settings';
 
-import '../stylesheets/SidebarContainer.scss';
-import { ApplicationContext } from '../context/ApplicationContext';
-import { AwsContext } from '../context/AwsContext';
-import { DashboardContext } from '../context/DashboardContext';
+import './styles.scss';
+import { ApplicationContext } from '../../context/ApplicationContext';
+import { AwsContext } from '../../context/AwsContext';
+import { DashboardContext } from '../../context/DashboardContext';
 
 const SidebarContainer = React.memo(() => {
-
   // Extract invervalID from ApplicationContext. Initival value: intervalID = null
-  const { intervalID,setExample } = useContext(ApplicationContext);
+  const { intervalID, setExample } = useContext(ApplicationContext);
   // Extract isLoading and setLoading state from AwsContext. Initial value: isLoading = true
   const { isLoading, setLoadingState } = useContext(AwsContext);
   // clear interval and set loading state to true when leaving graph containers
-  const { addApp } = useContext(DashboardContext)
+  const { addApp } = useContext(DashboardContext);
   /**
    * @function handleCLick - check if the 'intervalID' exists. If so, theres a timer running and the fuunction clears the timer using @function clearInterval - function.
    * Checks if variable 'isLoading' is false and if so the content is not loading and therefore, sets it to true using the setLoadingState function.
@@ -29,18 +28,17 @@ const SidebarContainer = React.memo(() => {
   };
 
   const handleExample = () => {
+    setExample(true);
 
-    setExample(true)
-    
     const fields = {
       typeOfService: 'Microservices',
       database: 'MongoDB',
-      URI: "mongodb://127.0.0.1:27017/?directConnection=true&serverSelectionTimeoutMS=2000&appName=mongosh+2.2.5",
+      URI: 'mongodb://127.0.0.1:27017/?directConnection=true&serverSelectionTimeoutMS=2000&appName=mongosh+2.2.5',
       name: 'Example',
-      description: 'Microservices'
-    }
-    addApp(fields)
-  }
+      description: 'Microservices',
+    };
+    addApp(fields);
+  };
 
   return (
     <div className="sidebar-container" id="mySidebar">
@@ -101,14 +99,10 @@ const SidebarContainer = React.memo(() => {
             &emsp;Contact
           </Link>
           <Link className="sidebar-link" to="/" id="dash" onClick={handleClick}>
-            <button 
-              className="example-button"
-              onClick={() => handleExample()}
-            >
-                EXAMPLE
+            <button className="example-button" onClick={() => handleExample()}>
+              EXAMPLE
             </button>
           </Link>
-
         </div>
       </div>
     </div>
