+++ app/modals/ServicesModal/ServicesModal.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .servicesContainer {
   width: 100px;
