+++ app/modals/UserModal/UserModal.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/constants.scss';
+@import '../../stylesheets/index.scss';
 
 .add-container {
   @include centerModal(30%, 500px);
