+++ app/stylesheets/Home.scss
@@ -1,147 +1,147 @@
-@import '../../stylesheets/constants.scss';
+// @import '../../stylesheets/constants.scss';
 
-.btns {
-  display: flex;
-  justify-content: space-around;
-}
+// .btns {
+//   display: flex;
+//   justify-content: space-around;
+// }
 
-.home {
-  background: $darkblue;
-  display: flex;
-  flex-direction: column;
-  align-content: center;
-  align-items: center;
-  justify-content: flex-start;
-  justify-items: center;
-  position: absolute;
-  margin-left: 50px;
-  padding-top: 160px;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  opacity: 1;
+// .home {
+//   background: $darkblue;
+//   display: flex;
+//   flex-direction: column;
+//   align-content: center;
+//   align-items: center;
+//   justify-content: flex-start;
+//   justify-items: center;
+//   position: absolute;
+//   margin-left: 50px;
+//   padding-top: 160px;
+//   top: 0;
+//   left: 0;
+//   width: 100vw;
+//   height: 100vh;
+//   opacity: 1;
 
 
-  .logo-container {
-    height: auto;
-    width: 40%;
-    max-width: 650px;
+//   .logo-container {
+//     height: auto;
+//     width: 40%;
+//     max-width: 650px;
 
-    img {
-      width: 100%;
-      transition: 0.3s ease-in;
+//     img {
+//       width: 100%;
+//       transition: 0.3s ease-in;
 
-      &:hover {
-        filter: brightness(0.7);
-      }
-    }
-  }
+//       &:hover {
+//         filter: brightness(0.7);
+//       }
+//     }
+//   }
 
-  h1 {
-    font-size: 325%;
-    letter-spacing: 2px;
-  }
+//   h1 {
+//     font-size: 325%;
+//     letter-spacing: 2px;
+//   }
 
-  p {
-    margin: 15px 0;
-    color: white;
-  }
+//   p {
+//     margin: 15px 0;
+//     color: white;
+//   }
 
-  // Get Started Button
-  .link {
-    font-size: 88%;
-    padding: 10px 17px;
-    letter-spacing: 0.5px;
-    text-decoration: none;
-    border-radius: 30px;
-    background-color: $navy;
-    color: #fff;
-    @include centerWithFlex();
-    transition: 1s;
-    animation: glow linear 4s infinite;
+//   // Get Started Button
+//   .link {
+//     font-size: 88%;
+//     padding: 10px 17px;
+//     letter-spacing: 0.5px;
+//     text-decoration: none;
+//     border-radius: 30px;
+//     background-color: $navy;
+//     color: #fff;
+//     @include centerWithFlex();
+//     transition: 1s;
+//     animation: glow linear 4s infinite;
 
-    &:hover {
-      background-color: $hover;
-      opacity: 0.85;
-      animation: none;
-    }
-  }
-}
+//     &:hover {
+//       background-color: $hover;
+//       opacity: 0.85;
+//       animation: none;
+//     }
+//   }
+// }
 
-.welcome {
-  color: rgba(255, 255, 255, 0.801);
-  text-align: center;
-  h1 {
-    margin-bottom: 90px;
-  }
-  h2 {
-    margin-bottom: 33px;
-  }
-  .btns {
-    display: flex;
-    justify-content: space-around;
-  }
-  form {
-    flex-direction: column;
-    label {
-      width: 50%;
-      input {
-        padding: 5px 7px;
-        margin: 2px 0;
-        width: 100%;
-      }
-    }
-    button.needAccount {
-      all: initial;
-      font-family: 'Roboto';
-      color: rgba(255, 255, 255, 0.801);
-      cursor: pointer;
-    }
-  }
-}
+// .welcome {
+//   color: rgba(255, 255, 255, 0.801);
+//   text-align: center;
+//   h1 {
+//     margin-bottom: 90px;
+//   }
+//   h2 {
+//     margin-bottom: 33px;
+//   }
+//   .btns {
+//     display: flex;
+//     justify-content: space-around;
+//   }
+//   form {
+//     flex-direction: column;
+//     label {
+//       width: 50%;
+//       input {
+//         padding: 5px 7px;
+//         margin: 2px 0;
+//         width: 100%;
+//       }
+//     }
+//     button.needAccount {
+//       all: initial;
+//       font-family: 'Roboto';
+//       color: rgba(255, 255, 255, 0.801);
+//       cursor: pointer;
+//     }
+//   }
+// }
 
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
+// @keyframes fadeIn {
+//   0% {
+//     opacity: 0;
+//   }
+//   100% {
+//     opacity: 1;
+//   }
+// }
 
-@-moz-keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
+// @-moz-keyframes fadeIn {
+//   0% {
+//     opacity: 0;
+//   }
+//   100% {
+//     opacity: 1;
+//   }
+// }
 
-@-webkit-keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
+// @-webkit-keyframes fadeIn {
+//   0% {
+//     opacity: 0;
+//   }
+//   100% {
+//     opacity: 1;
+//   }
+// }
 
-@-o-keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
+// @-o-keyframes fadeIn {
+//   0% {
+//     opacity: 0;
+//   }
+//   100% {
+//     opacity: 1;
+//   }
+// }
 
-@-ms-keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
+// @-ms-keyframes fadeIn {
+//   0% {
+//     opacity: 0;
+//   }
+//   100% {
+//     opacity: 1;
+//   }
+// }
