+++ app/stylesheets/scrollBar.scss
@@ -1,34 +1,34 @@
-/* width */
-::-webkit-scrollbar {
-  width: 10px;
-}
+// /* width */
+// ::-webkit-scrollbar {
+//   width: 10px;
+// }
 
-/* Track */
-::-webkit-scrollbar-track {
-  background: #f1f1f1;
-}
+// /* Track */
+// ::-webkit-scrollbar-track {
+//   background: #f1f1f1;
+// }
 
-/* Handle */
-::-webkit-scrollbar-thumb {
-  background: #888;
-}
+// /* Handle */
+// ::-webkit-scrollbar-thumb {
+//   background: #888;
+// }
 
-/* Handle on hover */
-::-webkit-scrollbar-thumb:hover {
-  background: #555;
-}
+// /* Handle on hover */
+// ::-webkit-scrollbar-thumb:hover {
+//   background: #555;
+// }
 
-body {
-  height: 100vh;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  #app {
-    height: 100vh;
-    overflow-y: scroll;
-    overflow-x: hidden;
-    &::-webkit-scrollbar {
-      width: 10px;
-    }
-  }
-}
+// body {
+//   height: 100vh;
+//   &::-webkit-scrollbar {
+//     display: none;
+//   }
+//   #app {
+//     height: 100vh;
+//     overflow-y: scroll;
+//     overflow-x: hidden;
+//     &::-webkit-scrollbar {
+//       width: 10px;
+//     }
+//   }
+// }
